separate time-series editor settings component
